fix: routing and refactor code
