+++ templates/login.html
@@ -7,7 +7,7 @@
 
         <div class="flex items-center justify-between">
             <h1 class="text-3xl font-bold mb-6">Log In</h1>
-            <a href="{{ url_for('index') }}" onclick="navigate(event, '/')" class="block mb-4">
+            <a href="{{ url_for('routes.index') }}" onclick="navigate(event, '/')" class="block mb-4">
                 <button type="button"
                     class="shadow bg-blue-500 text-white rounded-full  transition-colors hover:bg-blue-600">
                     <svg class="h-10 w-10 p-1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" fill="none"
@@ -45,7 +45,7 @@
             </div>
         </form>
 
-        <p class="mt-4 text-center">Don't have an account? <a href="{{ url_for('signup') }}"
+        <p class="mt-4 text-center">Don't have an account? <a href="{{ url_for('routes.signup') }}"
                 onclick="navigate(event, '/signup')" class="text-blue-500 hover:underline">Sign Up</a></p>
     </div>
 </div>
