+++ templates/update_chatbot.html
@@ -6,7 +6,7 @@
         class="p-5 w-full dark:text-dark border border-lighter dark:border-darker dark:bg-dark bg-light rounded-lg shadow drop-shadow">
         <div class="flex items-center justify-between">
             <h1 class="text-3xl font-bold mb-6">Update Chatbot</h1>
-            <a href="{{ url_for('dashboard') }}" onclick="navigate(event, '/dashboard')" class="block mb-4">
+            <a href="{{ url_for('routes.dashboard') }}" onclick="navigate(event, '/dashboard')" class="block mb-4">
                 <button type="button"
                     class="shadow bg-blue-500 text-white rounded-full  transition-colors hover:bg-blue-600">
                     <svg class="h-10 w-10 p-1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" fill="none"
